Added explicit type casting when adding coords to spacial objects. This sanatizes the data and allows the conversion to WKT to be used directly in queries without worry

+++ protected/components/spacial/SpacialPoint.php
@@ -35,8 +35,8 @@ class SpacialPoint extends SpacialAbstract
 		if (count($coord) != 2)
 			throw new Exception("Spacial Point coordinate requires two values. ".(count($coord))." found.");
 
-		$lat = isset($coord['lat']) ? $coord['lat'] : $coord[0];
-		$lon = isset($coord['lon']) ? $coord['lon'] : $coord[1];
+		$lat = (double)(isset($coord['lat']) ? $coord['lat'] : $coord[0]);
+		$lon = (double)(isset($coord['lon']) ? $coord['lon'] : $coord[1]);
 		$this->coords[] = compact('lat', 'lon');
 	}
 }
